Trying to implement the light model phong blinn
The implementation of phong-blinn continues.
Added 2 tests for visual verification:
- Get image from ptm, which is working
- Shade image in the ptm. Test is working but shading is not.

For now the problem is most probably, about normalizing the proper
thing, the equation should not have a problem, but at this point
I am not sure about anything. I'll wait for a Stack Overflow response
and see what i can come up with from there