Notes
So far most of the parameters are not very active.
I should concentrate on the use of costheta in order to make sure that light
direction affects the illumination of different parts of the tablet.
Then add other parameters one step at a time to know what they are doing