Opengl interface almost done
New window is added with a toolbox:
One page for file list, another page for viewer related options.
If things get to complicated I can move viewer options into two pages.

The rendering is now entirely done in opengl.
The normals are needed to be transformed into a texture, ie, QImage
before passing it to glwidget for displaying.

The events needed to be added but it should not take too long to wire
events