Lot of bugs in glwidget ptmparser needs tests
GL widget initialization does not work.
There is is problem with argument types for
vertexattribarray which is weird because
the same arguments work for light tutorial
which is more or less same setting.

I found out what to do with the normals of
ptm. Ptm file contains 3 normals
a normal map for each channel
I will simply add 2 other normals to shaders
to accomodate their effect.

1 red normal
1 green normal
1 blue normal

The shape of the normal array obtained
from coefficients also permit this.