New style texture buffer handling
Qt style handling of vertex buffer and attributes
seem to shield type errors that are used to be generated
I shall explore that a little more.
Instead of putting everything in a vertex buffer and trying
to bind the vertex array to that.
I shall use the setAttributeArray to specify values of each
attributes directly. This would render a big part of
renderQuadTriangle meaningless, and should give me a boost
in refactoring as well.