Normal map version renders with bugs
Good thing is that normal map version renders without error. But it is
rendered incorrectly. There is an issue with triangles I think.

The lightening seems like a phong with no ambient term.
It needs to be worked out. Even with the normal map it does not look like
a real light cast.
I should also fix the minor bugs like opening the rendering window in
new window instead of inside the application.
Simplest method is to place the widget exactly to the same place as
its predecessor. I should simply take the index of the viewerWidget
before initialization with ptm from parent and replace the initialized
widget