Started refactoring and new interface
Started a refactoring on the old interface.
I realized that the main issue is not with shaders but with the usage
of the ptm viewer in question. Instead of trying to write better
shaders I will concentrate on a more smooth interface with different
functionalities.
To that end, I switched to a docking widget instead of having control
widgets available all at once.
Three things are important:

- Reproductible light conditions. This is going to be possible
using serialization of lighting equation parameters.

- Having a capacity to take notes about the scene. This going to
be provided by the transcription dock widget.

- And of cource changing light conditions and camera. There
are several dock widgets dealing with functions related to these.

For the last part I should concentrate on moving light and camera
in every direction and rotating them at every direction as well.
I should implement the roll parameter for obtaining the view matrix
for the camera as well.

For now I will implement the functionalities using the per channel normal
map shader. I will disable the rest of it.

I should also consider and io for parameters. Outputing parameters in json
format is done, I might also consider an input specification for specifying
directly lightening parameters.