Changed inheritance structure.
The inheritance structure changed now.
In order to provide the same interface in pure and qt
implementations, qt and pure classes now inherit
from an abstract class. This abstract class
is written for camera, 3d object, and light.
The camera and 3d object are done.
The implementation of light with this new background
is on the way. The pure light source is implemented.
The QtLight source is not yet implemented.